+++ src/commands/hardis/work/refresh.ts
@@ -199,7 +199,7 @@ export default class RefreshTask extends SfdxCommand {
     }
 
     // Push new branch state to scratch org
-    await forceSourcePush(this.org.getUsername(), this, this.debugMode);
+    await forceSourcePush(this.org.getUsername(), this, this.debugMode, { conn: this.org.getConnection() });
 
     // Return an object to be displayed with --json
     return { outputString: "Refreshed the task & org" };