+++ src/common/gitProvider/github.ts
@@ -1,0 +1,11 @@
+import { GitProviderRoot } from "./gitProviderRoot";
+
+export class GithubProvider extends GitProviderRoot {
+  constructor() {
+    super();
+  }
+
+  public getLabel(): string {
+    return "sfdx-hardis GitHub connector";
+  }
+}